editing the price to display to 2decimals

+++ toy/views.py
@@ -63,7 +63,7 @@ def search(request):
 def one_toy(request, toy_id):
     toy = get_object_or_404(Toy, pk=toy_id)
 
-    toy.price = toy.price/100
+    toy.price = float(toy.price/100)
 
     return render(request, 'toy/one_toy.template.html', {
         'toy': toy
